Remove misleading TA properties
Some of our example TAs define the following properties:
 - "gp.ta.description" (type: string)
 - "gp.ta.version" (type: unsigned 32-bit integer)

This is misleading, because these are *not* the standard description and
version properties defined by the GP Internal Core API specification. The
standard ones are:
 - "gpd.ta.description" (type: string)
 - "gpd.ta.version" (type: string)
OP-TEE TAs should use the TA_DESCRIPTION and TA_VERSION macros to define
them.

This commit gets rid of the misleading properties. Instead, TA_DESCRIPTION
and TA_VERSION are set in the hello_world application. To still illustrate
the mechanism used to define non-standard properties, we add a couple of
values with names that follow the recommended reverse domain name
convention. Other TAs are left with no additional properties and default
version and description for simplicity.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>