Remove misleading TA properties in new added examples
Some of our example TAs define the following properties:
 - "gp.ta.description" (type: string)
 - "gp.ta.version" (type: unsigned 32-bit integer)

This is misleading, because these are *not* the standard description and
version properties defined by the GP Internal Core API specification. The
standard ones are:
 - "gpd.ta.description" (type: string)
 - "gpd.ta.version" (type: string)
OP-TEE TAs should use the TA_DESCRIPTION and TA_VERSION macros to define
them.

This commit gets rid of the misleading properties. Instead, TA_DESCRIPTION
and TA_VERSION are set in the applications.

Signed-off-by: ivila <390810839@qq.com>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@foss.st.com>