sign_verify: fix memory release issues
Remove memory release of the session context if failing to generate keys
in which case session closure, that is expected to free the session,
will trigger a segmentation fault.

For consistency of the sign_verify TA implementation, do not assume that
the session is closed each time after a command is invoked, hence free
transient objects once not more needed instead of at session closure.

Fixes: 911076b8e5ca ("sign_verify: Add dynamic algorithm selection with sign and verify support")
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>